requirements.txt
fastapi
uvicorn
pydantic
aiohttp
aiofiles
redis
mecab-python3
google-api-python-client
konlpy
mangum
requests==2.31.0
sentence-transformers==2.5.1
transformers==4.38.2
torch
urllib3
gliner==0.1.12
python-mecab-ko==1.3.5
python-dotenv
confluent-kafka
mysql-connector-python
huggingface-hub==0.23.0
pandas
aioredis==1.3.0
PyMySQL==1.1.1
cryptography==43.0.0
opensearch-py==2.6.0
